refactor: modularize page builder layout state
