+++ packages/ui/src/components/cms/page-builder/state/layout/duplicate.ts
@@ -1,7 +1,7 @@
 import type { HistoryState } from "@acme/types";
 import type { DuplicateAction } from "./types";
 import { commit } from "../history";
-import { duplicateComponent } from "./helpers";
+import { duplicateComponent } from "./utils";
 
 export function duplicate(state: HistoryState, action: DuplicateAction): HistoryState {
   return commit(state, duplicateComponent(state.present, action.id));
