+++ packages/ui/src/components/cms/page-builder/state/layout/move.ts
@@ -1,7 +1,7 @@
 import type { HistoryState } from "@acme/types";
 import type { MoveAction } from "./types";
 import { commit } from "../history";
-import { moveComponent } from "./helpers";
+import { moveComponent } from "./utils";
 
 export function move(state: HistoryState, action: MoveAction): HistoryState {
   return commit(state, moveComponent(state.present, action.from, action.to));
