+++ packages/ui/src/components/cms/page-builder/state/layout/update.ts
@@ -1,7 +1,7 @@
 import type { HistoryState } from "@acme/types";
 import type { UpdateAction } from "./types";
 import { commit } from "../history";
-import { updateComponent } from "./helpers";
+import { updateComponent } from "./utils";
 
 export function update(state: HistoryState, action: UpdateAction): HistoryState {
   return commit(state, updateComponent(state.present, action.id, action.patch));
