+++ packages/ui/src/components/cms/page-builder/state/reducer.d.ts
@@ -1,4 +1,4 @@
 import type { HistoryState } from "@acme/types";
-import type { Action } from "./actions";
+import type { Action } from "./layout";
 export declare function reducer(state: HistoryState, action: Action): HistoryState;
 export default reducer;
