+++ packages/ui/src/components/cms/page-builder/useCanvasResize.d.ts
@@ -1,4 +1,4 @@
-import type { ResizeAction } from "./state/actions";
+import type { ResizeAction } from "./state/layout";
 interface Options {
     componentId: string;
     widthKey: string;
